Handle large file uploads
The s3 client we're using chokes on large files. Plus, some repos will
have the same large files repeated across refs. So, this commit starts
handling large file uploads differently:

1. Identify large (> 500kiB) files and get their sha512 hash
2. Group files by content hash
3. Upload one copy of each hash to a private s3 directory
4. Delete large files from assembled directories
5. Upload assembled directories as normal
6. Copy large files in s3 from private blob to public locations

This greatly reduces upload time, especially since S3 is really, really
good at copying files that are already there.

+++ packages/assembler/src/providers/github-provider.js
@@ -22,6 +22,7 @@ const log = require('winston');
 const yaml = require('node-yaml');
 const moment = require('moment-timezone');
 const runCommand = require('../util/run-command');
+const path = require('path');
 
 const graphql = require('graphql.js');
 
@@ -156,12 +157,22 @@ module.exports = class GithubProvider {
     }
 
     async downloadRef(ref, destination) {
-        await fs.emptyDir(destination);
+        const exists = await fs.pathExists(path.join(destination, '.git'));
+        if (exists) {
+            await fs.ensureDir(destination);
+        } else {
+            await fs.emptyDir(destination);
+        }
 
         log.info(`Downloading ${this.source}@${ref} to ${destination} using git`);
 
         try {
-            await runCommand('git clone ' + this.source, 'git', ['clone', '--branch', ref, '--depth', '1', `https://${ghUser}:${ghToken}@github.com/${this.owner}/${this.repo}.git`, destination]);
+            if (exists) {
+                // await runCommand('git fetch ' + this.source, 'git', ['fetch', 'origin', ref]);
+                // await runCommand('git rebase ' + this.source, 'git', ['pull']);
+            } else {
+                await runCommand('git clone ' + this.source, 'git', ['clone', '--branch', ref, '--depth', '1', `https://${ghUser}:${ghToken}@github.com/${this.owner}/${this.repo}.git`, destination]);
+            }
         } catch (err) {
             log.error(`Error getting ${this.source}@${ref} via git`, err);
             throw err;
